More efficent planet generation and better camera speed control

+++ src/com/troy/ps/gamestate/GameState.java
@@ -8,7 +8,7 @@ public interface GameState {
 	
 	public void render(Window window);
 	
-	public void update(double delta, Window window);
+	public void update(float delta, Window window);
 	
 	public void cleanUp();
 }
